+++ internal/engine/evaluator.go
@@ -10,7 +10,6 @@ import (
 	"strings"
 
 	"github.com/google/cel-go/cel"
-	"github.com/google/cel-go/common/types"
 	"go.uber.org/multierr"
 	exprpb "google.golang.org/genproto/googleapis/api/expr/v1alpha1"
 	"google.golang.org/protobuf/types/known/emptypb"
@@ -58,212 +57,6 @@ func (noopEvaluator) Evaluate(_ context.Context, _ *enginev1.CheckInput) (*Polic
 type resourcePolicyEvaluator struct {
 	policy *runtimev1.RunnableResourcePolicySet
 	*tracer
-}
-
-type (
-	qpN   = enginev1.ResourcesQueryPlanOutput_Node
-	qpNLO = enginev1.ResourcesQueryPlanOutput_Node_LogicalOperation
-	qpNE  = enginev1.ResourcesQueryPlanOutput_Node_Expression
-	rN    = struct {
-		role string
-		f    func() (*qpN, error)
-		node *qpN
-	}
-)
-
-func (rpe *resourcePolicyEvaluator) EvaluateResourcesQueryPlan(_ context.Context, input *requestv1.ResourcesQueryPlanRequest) (*enginev1.ResourcesQueryPlanOutput, error) {
-	effectiveRoles := toSet(input.Principal.Roles)
-	inputActions := []string{input.Action}
-	result := &enginev1.ResourcesQueryPlanOutput{}
-	result.RequestId = input.RequestId
-	result.Kind = input.ResourceKind
-	result.Action = input.Action
-
-	for _, p := range rpe.policy.Policies { // zero or one policy in the set
-		var derivedRoles []rN
-
-		for drName, dr := range p.DerivedRoles {
-			dr := dr
-			if !setIntersects(dr.ParentRoles, effectiveRoles) {
-				continue
-			}
-
-			derivedRoles = append(derivedRoles, rN{
-				role: drName,
-				f: func() (*qpN, error) {
-					drVariables := make(map[string]*exprpb.Expr, len(dr.Variables))
-					for k, v := range dr.Variables {
-						drVariables[k] = v.Checked.Expr
-					}
-					node, err := evaluateCondition(dr.Condition, input, drVariables)
-					if err != nil {
-						return nil, err
-					}
-					return node, nil
-				},
-				node: nil,
-			})
-		}
-
-		// evaluate each rule until all actions have a result
-		for _, rule := range p.Rules {
-			var drNode *qpN
-			if !setIntersects(rule.Roles, effectiveRoles) {
-				nodes, err := getDerivedRoleConditions(derivedRoles, rule)
-				if err != nil {
-					return nil, err
-				}
-				switch len(nodes) {
-				case 0:
-					continue
-				case 1:
-					drNode = nodes[0]
-				default:
-					// combine restrictions (with OR) imposed by derived roles
-					drNode = &qpN{Node: &qpNLO{LogicalOperation: mkOrLogicalOperation(nodes)}}
-				}
-			}
-			if rule.Effect == effectv1.Effect_EFFECT_DENY {
-				panic("rules with effect DENY not supported")
-			}
-			for actionGlob := range rule.Actions {
-				matchedActions := globMatch(actionGlob, inputActions)
-				if len(matchedActions) == 0 {
-					continue
-				}
-				variables := make(map[string]*exprpb.Expr, len(p.Variables))
-				for k, v := range p.Variables {
-					variables[k] = v.Checked.Expr
-				}
-				var node *qpN
-				var err error
-				if rule.Condition != nil {
-					node, err = evaluateCondition(rule.Condition, input, variables)
-					if err != nil {
-						return nil, err
-					}
-				}
-				if drNode == nil {
-					result.Filter = node
-				} else {
-					// node AND drNode
-					if node != nil {
-						result.Filter = &qpN{Node: &qpNLO{LogicalOperation: mkAndLogicalOperation([]*qpN{drNode, node})}}
-					} else {
-						result.Filter = drNode
-					}
-				}
-			}
-		}
-	}
-	return result, nil
-}
-
-func getDerivedRoleConditions(derivedRoles []rN, rule *runtimev1.RunnableResourcePolicySet_Policy_Rule) ([]*qpN, error) {
-	var nodes []*qpN
-	for _, n := range derivedRoles {
-		if _, ok := rule.DerivedRoles[n.role]; ok {
-			node := n.node
-			var err error
-			if node == nil {
-				node, err = n.f()
-				if err != nil {
-					return nil, err
-				}
-				n.node = node
-			}
-			nodes = append(nodes, node)
-		}
-	}
-	return nodes, nil
-}
-
-func isNodeConstBool(node *enginev1.ResourcesQueryPlanOutput_Node) (bool, bool) {
-	if e, ok := node.Node.(*enginev1.ResourcesQueryPlanOutput_Node_Expression); ok {
-		if e1 := e.Expression.GetExpr().GetConstExpr(); e1 != nil {
-			if b, ok := e1.ConstantKind.(*exprpb.Constant_BoolValue); ok {
-				return b.BoolValue, true
-			}
-		}
-	}
-
-	return false, false
-}
-
-func mkOrLogicalOperation(nodes []*enginev1.ResourcesQueryPlanOutput_Node) *enginev1.ResourcesQueryPlanOutput_LogicalOperation {
-	return &enginev1.ResourcesQueryPlanOutput_LogicalOperation{
-		Operator: enginev1.ResourcesQueryPlanOutput_LogicalOperation_OPERATOR_OR,
-		Nodes:    nodes,
-	}
-}
-
-func mkAndLogicalOperation(nodes []*enginev1.ResourcesQueryPlanOutput_Node) *enginev1.ResourcesQueryPlanOutput_LogicalOperation {
-	return &enginev1.ResourcesQueryPlanOutput_LogicalOperation{
-		Operator: enginev1.ResourcesQueryPlanOutput_LogicalOperation_OPERATOR_AND,
-		Nodes:    nodes,
-	}
-}
-
-func evaluateCondition(condition *runtimev1.Condition, input *requestv1.ResourcesQueryPlanRequest, variables map[string]*exprpb.Expr) (*enginev1.ResourcesQueryPlanOutput_Node, error) {
-	res := new(qpN)
-	switch t := condition.Op.(type) {
-	case *runtimev1.Condition_Any:
-		nodes := make([]*qpN, 0, len(t.Any.Expr))
-		for _, c := range t.Any.Expr {
-			node, err := evaluateCondition(c, input, variables)
-			if err != nil {
-				return nil, err
-			}
-
-			add := true
-
-			if b, ok := isNodeConstBool(node); ok {
-				if b {
-					res.Node = &qpNE{Expression: conditions.TrueExpr}
-					return res, nil
-				}
-				add = false
-			}
-
-			if add {
-				nodes = append(nodes, node)
-			}
-		}
-		res.Node = &qpNLO{
-			LogicalOperation: mkOrLogicalOperation(nodes),
-		}
-	case *runtimev1.Condition_All:
-		nodes := make([]*qpN, 0, len(t.All.Expr))
-		for _, c := range t.All.Expr {
-			node, err := evaluateCondition(c, input, variables)
-			if err != nil {
-				return nil, err
-			}
-			add := true
-
-			if b, ok := isNodeConstBool(node); ok {
-				if !b {
-					res.Node = &qpNE{Expression: conditions.FalseExpr}
-					return res, nil
-				}
-				add = false
-			}
-
-			if add {
-				nodes = append(nodes, node)
-			}
-		}
-		res.Node = &qpNLO{LogicalOperation: mkAndLogicalOperation(nodes)}
-	case *runtimev1.Condition_Expr:
-		_, residual, err := evaluateCELExprPartially(t.Expr.Checked, input, variables)
-		if err != nil {
-			return nil, fmt.Errorf("error evaluating condition %q: %w", t.Expr.Original, err)
-		}
-		res.Node = &qpNE{Expression: residual}
-	default:
-		return nil, fmt.Errorf("unsupported condition type %T", t)
-	}
-	return res, nil
 }
 
 func (rpe *resourcePolicyEvaluator) Evaluate(ctx context.Context, input *enginev1.CheckInput) (*PolicyEvalResult, error) {
@@ -517,55 +310,6 @@ func evaluateBoolCELExpr(expr *exprpb.CheckedExpr, variables map[string]interfac
 	return boolVal, nil
 }
 
-func evaluateCELExprPartially(expr *exprpb.CheckedExpr, input *requestv1.ResourcesQueryPlanRequest, variables map[string]*exprpb.Expr) (*bool, *exprpb.CheckedExpr, error) {
-	e := expr.Expr
-	err := replaceVars(&e, variables)
-	if err != nil {
-		return nil, nil, err
-	}
-	ast := cel.ParsedExprToAst(&exprpb.ParsedExpr{Expr: e})
-	env := conditions.StdPartialEnv
-	prg, err := env.Program(ast, cel.EvalOptions(cel.OptPartialEval, cel.OptTrackState))
-	if err != nil {
-		return nil, nil, err
-	}
-
-	vars, err := cel.PartialVars(map[string]interface{}{
-		conditions.CELRequestIdent:                   input,
-		conditions.CELPrincipalAbbrev:                input.Principal,
-		conditions.Fqn(conditions.CELPrincipalField): input.Principal,
-	},
-		cel.AttributePattern(conditions.CELResourceAbbrev),
-		cel.AttributePattern(conditions.CELRequestIdent).QualString(conditions.CELResourceField))
-	if err != nil {
-		return nil, nil, err
-	}
-
-	val, details, err := prg.Eval(vars)
-	if err != nil {
-		return nil, nil, err
-	}
-	residual, err := env.ResidualAst(ast, details)
-	if err != nil {
-		return nil, nil, err
-	}
-	ast, iss := env.Check(residual)
-	if iss != nil {
-		return nil, nil, fmt.Errorf("failed to check residual express: %w", iss.Err())
-	}
-	checkedExpr, err := cel.AstToCheckedExpr(ast)
-	if err != nil {
-		return nil, nil, err
-	}
-	if types.IsUnknown(val) {
-		return nil, checkedExpr, nil
-	}
-	if b, ok := val.Value().(bool); ok {
-		return &b, checkedExpr, nil
-	}
-	return nil, checkedExpr, fmt.Errorf("unexpected result type %T", val.Value())
-}
-
 func evaluateCELExpr(expr *exprpb.CheckedExpr, variables map[string]interface{}, input *enginev1.CheckInput) (interface{}, error) {
 	if expr == nil {
 		return nil, nil
@@ -674,93 +418,4 @@ func (er *PolicyEvalResult) setDefaultEffect(tctx *traceContext, actions []strin
 			tctx.beginTrace(actionComponent, a).writeEvent(KVEffect(effect), KVMsg("Default effect"))
 		}
 	}
-}
-
-func updateIds(e *exprpb.Expr) {
-	var n int64
-
-	var impl func(e *exprpb.Expr)
-	impl = func(e *exprpb.Expr) {
-		if e == nil {
-			return
-		}
-		n++
-		e.Id = n
-		switch e := e.ExprKind.(type) {
-		case *exprpb.Expr_SelectExpr:
-			impl(e.SelectExpr.Operand)
-		case *exprpb.Expr_CallExpr:
-			impl(e.CallExpr.Target)
-			for _, arg := range e.CallExpr.Args {
-				impl(arg)
-			}
-		case *exprpb.Expr_StructExpr:
-			for _, entry := range e.StructExpr.Entries {
-				impl(entry.GetMapKey())
-				impl(entry.GetValue())
-			}
-		case *exprpb.Expr_ComprehensionExpr:
-			ce := e.ComprehensionExpr
-			impl(ce.IterRange)
-			impl(ce.AccuInit)
-			impl(ce.LoopStep)
-			impl(ce.LoopCondition)
-			impl(ce.Result)
-		case *exprpb.Expr_ListExpr:
-			for _, element := range e.ListExpr.Elements {
-				impl(element)
-			}
-		}
-	}
-	impl(e)
-}
-
-func replaceVars(e **exprpb.Expr, vars map[string]*exprpb.Expr) (err error) {
-	var r func(e *exprpb.Expr) *exprpb.Expr
-	r = func(e *exprpb.Expr) *exprpb.Expr {
-		if e == nil {
-			return nil
-		}
-		switch e := e.ExprKind.(type) {
-		case *exprpb.Expr_SelectExpr:
-			ident := e.SelectExpr.Operand.GetIdentExpr()
-			if ident != nil && (ident.Name == conditions.CELVariablesAbbrev || ident.Name == conditions.CELVariablesIdent) {
-				if v, ok := vars[e.SelectExpr.Field]; ok {
-					return v
-				}
-				err = multierr.Append(err, fmt.Errorf("unknown variable %q", e.SelectExpr.Field))
-			} else {
-				e.SelectExpr.Operand = r(e.SelectExpr.Operand)
-			}
-		case *exprpb.Expr_CallExpr:
-			e.CallExpr.Target = r(e.CallExpr.Target)
-			for i, arg := range e.CallExpr.Args {
-				e.CallExpr.Args[i] = r(arg)
-			}
-		case *exprpb.Expr_StructExpr:
-			for _, entry := range e.StructExpr.Entries {
-				if k, ok := entry.KeyKind.(*exprpb.Expr_CreateStruct_Entry_MapKey); ok {
-					k.MapKey = r(k.MapKey)
-				}
-				entry.Value = r(entry.Value)
-			}
-		case *exprpb.Expr_ComprehensionExpr:
-			ce := e.ComprehensionExpr
-			ce.IterRange = r(ce.IterRange)
-			ce.AccuInit = r(ce.AccuInit)
-			ce.LoopStep = r(ce.LoopStep)
-			ce.LoopCondition = r(ce.LoopCondition)
-			// ce.Result seems to be always an identifier, so isn't necessary to process
-		case *exprpb.Expr_ListExpr:
-			for i, element := range e.ListExpr.Elements {
-				e.ListExpr.Elements[i] = r(element)
-			}
-		}
-		return e
-	}
-
-	*e = r(*e)
-	updateIds(*e)
-
-	return err
 }
